refactor: split preprocess folder sweep into its own worker
ExtractionCleanupWorker had grown two unrelated cleanup domains with four
counter ivars and duplicated dry-run branching, tripping reek's
TooManyInstanceVariables/TooManyStatements/RepeatedConditional/FeatureEnvy
and regressing the quality gate. PreProcessCleanupWorker now owns the
preprocess sweep end to end (its own policy load, counters, dry-run
handling, and per-folder failure reporting), scheduled 30 minutes after
the extraction cleanup so the two nightly sweeps don't start together.
